[view-transitions] Visible clipping at the end of animation on https://www.kvin.me/posts/cards https://bugs.webkit.org/show_bug.cgi?id=274404 <rdar://128406756>
Reviewed by Alan Baradlay.

RenderLayer's use of RenderInline doesn't include any padding from its containing block in its position, and instead accounts for it
in the position of the lines boxes themselves.

This means that the computed (web visible) transform (using the renderer's position) for the ::view-transition-group doesn't
include the offset between the inline element and its container. It also means that an overflow:clip on the
::view-transition-image-pair applies the clip in the wrong coordinate space for the snapshots.

This change computes the offset between the position of the renderer, and that used by RenderLayer. It uses this
to adjust the computed transform, and the overflow rect of the snapshot so that they correctly account for the position of the
renderer.

* LayoutTests/imported/w3c/web-platform-tests/css/css-view-transitions/inline-with-offset-from-containing-block-clipped-ref.html: Added.
* LayoutTests/imported/w3c/web-platform-tests/css/css-view-transitions/new-content-inline-with-offset-from-containing-block-clipped-expected.html: Added.
* LayoutTests/imported/w3c/web-platform-tests/css/css-view-transitions/new-content-inline-with-offset-from-containing-block-clipped.html: Added.
* LayoutTests/imported/w3c/web-platform-tests/css/css-view-transitions/old-content-inline-with-offset-from-containing-block-clipped-expected.html: Added.
* LayoutTests/imported/w3c/web-platform-tests/css/css-view-transitions/old-content-inline-with-offset-from-containing-block-clipped.html: Added.
* Source/WebCore/dom/ViewTransition.cpp:
(WebCore::layerToLayoutOffset):
(WebCore::ViewTransition::captureOldState):
(WebCore::ViewTransition::copyElementBaseProperties):
(WebCore::ViewTransition::updatePseudoElementStyles):
* Source/WebCore/dom/ViewTransition.h:
* Source/WebCore/rendering/RenderViewTransitionCapture.cpp:
(WebCore::RenderViewTransitionCapture::setCapturedSize):
(WebCore::RenderViewTransitionCapture::layout):
(WebCore::RenderViewTransitionCapture::setSize): Deleted.
* Source/WebCore/rendering/RenderViewTransitionCapture.h:
* Source/WebCore/rendering/updating/RenderTreeUpdaterViewTransition.cpp:
(WebCore::createRendererIfNeeded):

Canonical link: https://commits.webkit.org/279433@main