Need to SUPPRESS_ASAN on MetadataTable::isDestroyed(). https://bugs.webkit.org/show_bug.cgi?id=274844 rdar://128875400
Reviewed by Yijia Huang.

MetadataTable::isDestroyed() is used to check if the unlinkedMetadata Ref is null
as a null check mitigation.  Under normal circumstances, when a Ref is leaked and
nullified, its memory gets poison on ASAN to ensure that it is not accessed
thereafter.  This conflicts with our mitigation which wishes to access it after it
is nullified.  The fix here is simply to apply SUPPRESS_ASAN to
MetadataTable::isDestroyed().

* Source/JavaScriptCore/bytecode/MetadataTable.h:
(JSC::MetadataTable::isDestroyed const):

Canonical link: https://commits.webkit.org/279451@main