[Monterey] Videos don't fully render in LinkedIn news feed https://bugs.webkit.org/show_bug.cgi?id=274812 rdar://127264745
Reviewed by Youenn Fablet.

Video frames were not presented as when a requestVideoFrameCallback operation is pending we do
not create the required AVSampleBufferDisplayLayer.
This is a regression from 252893@main.
On OS that do not support [AVSBDL copyDisplayedPixelBuffer] (Monterey and earlier), we create
a WebCoreDecompressionSession in place of a AVSBDL to decode the video frames. But those frames
are never rendered to screen. The (incorrect) assumption was that when using rVFC we would always
use a canvas or WebGL context to paint the video frames.
For now we revert the line introduced by 252893@main that while fixing rVFC broke playback.
rVFC was broken anyway on Monterey and earlier (you never get the expected frame).

No test, as we no longer have infrastructure in place to check on specific Monterey code path.

* Source/WebCore/platform/graphics/avfoundation/objc/MediaPlayerPrivateMediaSourceAVFObjC.mm:
(WebCore::MediaPlayerPrivateMediaSourceAVFObjC::shouldEnsureLayerOrVideoRenderer const):

Canonical link: https://commits.webkit.org/279474@main