[Unified PDF] pages are blank/white for 3-10s after scrolling quickly in 2-up https://bugs.webkit.org/show_bug.cgi?id=274862 rdar://128405116
Reviewed by Abrar Rahman Protyasha.

AsyncPDFRenderer was failing to keep track of already-requested page previews;
it never added to or removed from `m_enqueuedPagePreviews`. This caused it to
repeatedly decode previews for the same page.

Fix by adding to `m_enqueuedPagePreviews` in `generatePreviewImageForPage`
and removing in `didCompletePagePreviewRender()`, which is code factored
out of the main thread callback that was in `paintPagePreviewOnWorkQueue()`.

* Source/WebKit/WebProcess/Plugins/PDF/UnifiedPDF/AsyncPDFRenderer.h:
* Source/WebKit/WebProcess/Plugins/PDF/UnifiedPDF/AsyncPDFRenderer.mm:
(WebKit::AsyncPDFRenderer::generatePreviewImageForPage):
(WebKit::AsyncPDFRenderer::paintPagePreviewOnWorkQueue):
(WebKit::AsyncPDFRenderer::didCompletePagePreviewRender):

Canonical link: https://commits.webkit.org/279489@main