Speedometer 3: Fix the overflow:scroll layerization on TodoMVC-React-Complex-DOM. https://bugs.webkit.org/show_bug.cgi?id=268117 <rdar://113239854>
Reviewed by NOBODY (OOPS!).

This page has a composited layer that overlaps the right edge of scroller. Most of the
sub-layers inside the scroller are on the left edge and can't ever overlap the layer
outside the scroller (since the scroller only moves vertically). We currently can't detect
that there won't ever be overlap between the sublayers and the outside layer, so everything
in the scroller becomes composited too.

This adds support for overlapping backing provider candidates.

In order for a layer to be added to a candidate that isn't the topmost, we need to ensure
that the layer doesn't (and won't) intersect with the candidates above it.

Generally we just use the bounds of the candidate for this intersection test, which should
have the same results as disallowing overlapping candidates (except for having more in the
candidate list, and potentially worse performance checking them).

If the candidate is composited scrolling that only moves in the vertical direction, we use
the bounds of the current layer in the horizontal direction (and the bounds of the candidate
in the vertical).

This relies on the list of backing sharing candidates being sorted in z-order, which wasn't
the case previosuly (candidates were added after traversing their descendants). This now captures
a snapshot of the backing sharing state before traversing descendants, and then uses that to
insert it at the right spot after descendant traversal.

* LayoutTests/compositing/shared-backing/overflow-scroll/nested-scroller-sharing-expected.txt:
* LayoutTests/compositing/shared-backing/overflow-scroll/overlapping-scroller-sharing-expected.txt:
* LayoutTests/fast/repaint/background-attachment-fixed-in-composited-scroll-expected.txt:
* Source/WTF/Scripts/Preferences/UnifiedWebPreferences.yaml:
* Source/WebCore/rendering/RenderLayerCompositor.cpp:
(WebCore::RenderLayerCompositor::BackingSharingState::BackingSharingState):
(WebCore::RenderLayerCompositor::BackingSharingState::snapshot):
(WebCore::RenderLayerCompositor::BackingSharingState::startBackingSharingSequence):
(WebCore::RenderLayerCompositor::BackingSharingState::addBackingSharingCandidate):
(WebCore::RenderLayerCompositor::BackingSharingState::endBackingSharingSequence):
(WebCore::RenderLayerCompositor::BackingSharingState::backingProviderCandidateForLayer):
(WebCore::RenderLayerCompositor::BackingSharingState::isAdditionalProviderCandidate const):
(WebCore::RenderLayerCompositor::updateCompositingLayers):
(WebCore::RenderLayerCompositor::computeCompositingRequirements):
(WebCore::RenderLayerCompositor::traverseUnchangedSubtree):
(WebCore::RenderLayerCompositor::updateBackingSharingBeforeDescendantTraversal):
(WebCore::RenderLayerCompositor::updateBackingSharingAfterDescendantTraversal):
(WebCore::RenderLayerCompositor::computeClippedOverlapBounds const):
(WebCore::RenderLayerCompositor::addToOverlapMap const):
* Source/WebCore/rendering/RenderLayerCompositor.h: