[Site Isolation] Remove child frames of old main frame when committing a provisional page https://bugs.webkit.org/show_bug.cgi?id=274941 rdar://125374765
Reviewed by Charlie Wolfe.

When a main frame navigates to a new domain, we need the iframes of the previous page to be
removed in the UI process and in all other processes.  This isn't an issue with site isolation
off because we cache or discard the entirety of the previous process when process swapping.
With site isolation, though, the same processes can be reused for the same domains after
main frame navigations.

* Source/WebKit/UIProcess/WebFrameProxy.cpp:
(WebKit::WebFrameProxy::removeRemoteFrameProcessUse):
* Source/WebKit/UIProcess/WebFrameProxy.h:
* Source/WebKit/UIProcess/WebPageProxy.cpp:
(WebKit::WebPageProxy::commitProvisionalPage):
* Tools/TestWebKitAPI/Tests/WebKitCocoa/SiteIsolation.mm:
(TestWebKitAPI::TEST(SiteIsolation, NavigateToDomainAgainWithSPIUse)):

Canonical link: https://commits.webkit.org/279561@main