AX: empty strings shouldn't be cached in isolated tree https://bugs.webkit.org/show_bug.cgi?id=274392 rdar://problem/128397189
Reviewed by Tyler Wilcock.

AXIsolatedObject::setProperty checks to see if a property value is
equal to the default to avoid setting it. This logic has a small bug
when it comes to strings, because the null string and empty string are
not considered equal, but we don't need to store either in the
property map. By calling isEmpty() instead of comparing to
emptyString(), we can ensure no empty strings are stored in the
property map.

The impact of this change on memory usage is negligible by itself, but
it will help future optimizations by making the overall property map
significantly smaller.

* Source/WebCore/accessibility/isolatedtree/AXIsolatedObject.cpp:
(WebCore::AXIsolatedObject::setProperty):
(WebCore::AXIsolatedObject::stringAttributeValueNullIfMissing const):
* Source/WebCore/accessibility/isolatedtree/AXIsolatedObject.h:

Canonical link: https://commits.webkit.org/279669@main