Fix page load time regression related to Web Audio https://bugs.webkit.org/show_bug.cgi?id=275057 rdar://129185678
Reviewed by Ben Nham.

We are observing a page load time issue, where it appears to be beneficial to eagerly create an audio
destination in the GPU process. The reason that this improves page load time is likely that creating
an audio destination in the GPU process will prewarm audio related frameworks in the GPU proceess and
benefit subsequent page loads that use audio features. We also resolved the same issue in an earlier
patch. See rdar://104617724.

* Source/WebKit/WebProcess/GPU/media/RemoteAudioDestinationProxy.cpp:
(WebKit::RemoteAudioDestinationProxy::RemoteAudioDestinationProxy):
(WebKit::RemoteAudioDestinationProxy::stopRendering):

Canonical link: https://commits.webkit.org/279676@main