[UnifiedPDF] Cannot zoom out on certain large documents https://bugs.webkit.org/show_bug.cgi?id=275072 rdar://128915788
Reviewed by Simon Fraser.

Some clients may have a lower and upper bounds in which they restrict
their page scales. In certain circumstances we may compute a normalized
scale factor which ends up being below the upper bound. It appears that
this can occur frequently in large documents because we compute a
"large," scale for the actual size and end up computing a "small,"
normalization factor. This is what ends up happening if you open the PDF
in the bugzilla with Minibrowser. To limit this pathological behavior
with the increased document size let's make sure that when we compute
the normalization scale factor it is relative to the actual size scale.

* Source/WebKit/WebProcess/Plugins/PDF/UnifiedPDF/UnifiedPDFPlugin.mm:
(WebKit::UnifiedPDFPlugin::computeNormalizationFactor):

Canonical link: https://commits.webkit.org/279688@main