Get Upstream Commit reference for people using a local source https://bugs.webkit.org/show_bug.cgi?id=272142 rdar://125894887
Reviewed by Jonathan Bedard.

This provides the hash of the HEAD for people using a local checkout
of WPT tests. This is a followup on webkit.org/b/248445 which gave
this information for people using the remote checkout. We wrapped in
a try/except to make sure the directory was really a checkout of a repo.

It makes it possible for people to directly copy paste in the commit
message the hash of the WPT checkout when they import
or re-import WPT tests.

Example:
--------- Please include the following in your commit message: ---------
Upstream commit: https://github.com/web-platform-tests/wpt/commit/3324586879a38951098d072bf5ae8f56408a9afa
------------------------------------------------------------------------

* Tools/Scripts/webkitpy/w3c/test_downloader.py:
(TestDownloader.checkout_test_repository):
(TestDownloader):
(TestDownloader.get_rev_parse_head):
* Tools/Scripts/webkitpy/w3c/test_importer.py:
(TestImporter.do_import):

Canonical link: https://commits.webkit.org/279698@main