[WebGPU] copyExternalImageToTexture incorrectly copies images from HTMLImageElement https://bugs.webkit.org/show_bug.cgi?id=274789 <radar://128380266>
Reviewed by Dan Glastonbury.

We were assuming the byte order was always RGBA and not handling
ARGB or ABGR, where A can be X or A.

Correct this by manually swapping.

https://bugs.webkit.org/show_bug.cgi?id=263692 tracks moving this to the GPU
process and we could perform the swaps in a fragment shader as the goal is to
populate a MTLTexture.

Re-land with release assert in correct location.

* Source/WebCore/Modules/WebGPU/GPUQueue.cpp:
(WebCore::getImageBytesFromImageBuffer):
(WebCore::getImageBytesFromVideoFrame):
(WebCore::imageBytesForSource):
(WebCore::populdateXYFromOrigin):
(WebCore::copyToDestinationFormat):
(WebCore::GPUQueue::copyExternalImageToTexture):

Canonical link: https://commits.webkit.org/279714@main