Send all scroll events to the visualViewport https://bugs.webkit.org/show_bug.cgi?id=218465
Reviewed by NOBODY (OOPS!).

The previous code would send a single scroll event update to the
visual viewport interface when the visual viewport was updated, but
other browsers send all document scroll events to the visual viewport.

I opted to keep the needsVisualViewportScrollEvent boolean as a slight
optimization so that we don't send events if nothing is listening to
them.

* Source/WebCore/dom/Document.cpp:
(WebCore::Document::runScrollSteps): Do not clear the needs visual
viewport scroll event flag, so that we keep delivering scroll events to
the visualViewport.

* Source/WebCore/page/VisualViewport.cpp:
(WebCore::VisualViewport::addEventListener): Tell the document to send
scroll events to the visualViewport when a scroll listener is added.
(WebCore::VisualViewport::update): Simplify the logic now that this is
not responsible for deciding if scroll events should be sent.