GNOME Discourse page loads forever https://bugs.webkit.org/show_bug.cgi?id=274727
Reviewed by NOBODY (OOPS!).

The problem here is the load actually finishes, but PageLoadState
doesn't know it has finished because PageLoadState::isLoading always
returns true if there is a pending API request URL. It has to, because
otherwise it would return false immediately after starting a load,
because the default state is Finished.

Normally the pending API request gets unset before the load finishes,
but if it doesn't, then WebKit will get stuck in the isLoading state.

This fix doesn't feel particularly satisfying. We should probably also
add a test for this, but I don't know which conditions trigger the bug.
It requires restoring a page from session state, but that is not
sufficient.

* Source/WebKit/UIProcess/PageLoadState.cpp:
(WebKit::PageLoadState::didFinishProgress):