Log top origin of opener window when tracking cross-tab WindowProxy usage https://bugs.webkit.org/show_bug.cgi?id=274999 rdar://129085170
Reviewed by John Wilander.

When tracking cross-tab WindowProxy usage, currently we are logging the site of the opener frame
(which could be an iframe) and the site of the openee frame. Instead we want to log the site
associated with the opener's main frame, since what we care about is potential sharing of top-level
cookies between the tabs.

Additionally, we should suppress logging when there is cross-tab WindowProxy usage but the two tab's
main frames have the same site. The two tabs are already sharing top-level cookies in that case so
we don't need to log in that case.

Finally, we used to make the distinction between logging the actual property type being accessed
(postMessage vs. closed vs. other). But due to event logging constraints we probably don't want to
be doing this for now as it is tripling the event volume in many cases. So for now we just log every
property access in the "other"" category.

* LayoutTests/TestExpectations:
* LayoutTests/http/wpt/opener/cross-site-child-access-parent-iframe-via-windowproxy-expected.txt: Added.
* LayoutTests/http/wpt/opener/cross-site-child-access-parent-iframe-via-windowproxy.html: Added.
* LayoutTests/http/wpt/opener/resources/iframe-open-window.html: Added.
* LayoutTests/http/wpt/opener/resources/send-message-to-opener.html:
* LayoutTests/http/wpt/opener/same-site-child-access-parent-iframe-via-windowproxy-expected.txt: Added.
* LayoutTests/http/wpt/opener/same-site-child-access-parent-iframe-via-windowproxy.html: Added.
* LayoutTests/platform/ios/TestExpectations:
* LayoutTests/platform/mac-site-isolation/TestExpectations:
* LayoutTests/platform/mac-wk2/TestExpectations:
* Source/WebCore/page/LocalFrame.cpp:
(WebCore::LocalFrame::didAccessWindowProxyPropertyViaOpener):

Canonical link: https://commits.webkit.org/279783@main