After closing a modal <dialog>, elements can be z-ordered incorrectly https://bugs.webkit.org/show_bug.cgi?id=275188 rdar://128899243
Reviewed by Antti Koivisto.

When adding or removing an element from the top layer (e.g. for dialog or popups),
`RenderLayer::establishesTopLayerDidChange()` is called. This called `parentLayer->addChild()`
which re-parents the RenderLayer, but it always appended it to the end of the list. This would
cause adjacent `position:absolute` elements to z-order incorrectly when closing a modal dialog.

* LayoutTests/fast/layers/layer-order-after-top-layer-expected.html: Added.
* LayoutTests/fast/layers/layer-order-after-top-layer.html: Added.
* Source/WebCore/rendering/RenderLayer.cpp:
(WebCore::RenderLayer::establishesTopLayerDidChange):

Canonical link: https://commits.webkit.org/279785@main