[WebGPU] GPURenderPipeline can surprisingly be garbage collected https://bugs.webkit.org/show_bug.cgi?id=275194 <radar://129271876>
Reviewed by Dan Glastonbury.

We should hold a RefPtr and not a WeakPtr since a GC which occurs in the middle
of an encoder can result in the pipeline's ref count reaching zero.

There is not a risk of a retain cycle as the *Encoders are not retained by
any other objects.

* LayoutTests/fast/webgpu/regression/repro_275194-expected.txt: Added.
* LayoutTests/fast/webgpu/regression/repro_275194.html: Added.
* LayoutTests/fast/webgpu/regression/repro_275194b-expected.txt: Added.
* LayoutTests/fast/webgpu/regression/repro_275194b.html: Added.
Add regression test.

* Source/WebGPU/WebGPU/ComputePassEncoder.h:
* Source/WebGPU/WebGPU/RenderBundleEncoder.h:
* Source/WebGPU/WebGPU/RenderPassEncoder.h:

Canonical link: https://commits.webkit.org/279795@main