[WebGPU] GPUTextureView can surprisingly be garbage collected https://bugs.webkit.org/show_bug.cgi?id=275196 <radar://129272236>
Reviewed by Dan Glastonbury.

We should hold a RefPtr and not a WeakPtr since a GC which occurs in the middle
of an encoder can result in the pipeline's ref count reaching zero.

There is not a risk of a retain cycle as the *Encoders are not retained by
any other objects.

* LayoutTests/fast/webgpu/regression/repro_275196-expected.txt: Added.
* LayoutTests/fast/webgpu/regression/repro_275196.html: Added.
* LayoutTests/fast/webgpu/regression/repro_275196b-expected.txt: Added.
* LayoutTests/fast/webgpu/regression/repro_275196b.html: Added.
* Source/WebGPU/WebGPU/RenderBundle.h:
* Source/WebGPU/WebGPU/RenderBundle.mm:
(WebGPU::RenderBundle::validateRenderPass const):
* Source/WebGPU/WebGPU/RenderPassEncoder.h:
* Source/WebGPU/WebGPU/RenderPassEncoder.mm:
(WebGPU::m_maxDrawCount):
* Source/WebGPU/WebGPU/RenderPipeline.h:
* Source/WebGPU/WebGPU/RenderPipeline.mm:
(WebGPU::RenderPipeline::colorDepthStencilTargetsMatch const):

Canonical link: https://commits.webkit.org/279797@main